Move all softmask functions to a separate script
